refactor: move `GetTransaction(...)` to node/transaction.cpp
can be reviewed with --color-moved
